Rename `lib` to `source` and move files
