--- index.js
@@ -1,5 +1,0 @@
-export {getStreamAsArray} from './lib/array.js';
-export {getStreamAsArrayBuffer} from './lib/array-buffer.js';
-export {getStreamAsBuffer} from './lib/buffer.js';
-export {getStreamAsString as default} from './lib/string.js';
-export {MaxBufferError} from './lib/contents.js';
