+++ source/array-buffer.js
