+++ source/array.js
