+++ source/buffer.js
