+++ source/index.js
@@ -1,0 +1,5 @@
+export {getStreamAsArray} from './array.js';
+export {getStreamAsArrayBuffer} from './array-buffer.js';
+export {getStreamAsBuffer} from './buffer.js';
+export {getStreamAsString as default} from './string.js';
+export {MaxBufferError} from './contents.js';
