+++ source/string.js
