+++ test/string.js
@@ -1,7 +1,7 @@
 import {Buffer, constants as BufferConstants} from 'node:buffer';
 import {text} from 'node:stream/consumers';
 import test from 'ava';
-import getStream, {MaxBufferError} from '../index.js';
+import getStream, {MaxBufferError} from '../source/index.js';
 import {createStream, BIG_TEST_DURATION} from './helpers/index.js';
 import {
 	fixtureString,
